+++ pan_mcp_user_role/__manifest__.py
@@ -4,7 +4,7 @@
 
 {
     "name": "User Roles (Pantalytics bundle)",
-    "version": "19.0.1.0.4",
+    "version": "19.0.1.0.5",
     "category": "Tools",
     "author": "ABF OSIELL, Odoo Community Association (OCA)",
     "license": "LGPL-3",